backend/requirements_3.8_pro.txt
astroid==2.3.3
bleach==3.1.0
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.4.4
six==1.13.0
SQLAlchemy==1.3.11
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
